requirements.txt
future
emoji
lxml
wget
gtts
beautifulsoup4
requests==2.20.0
python-telegram-bot==13.7
sqlalchemy==1.3.23
psycopg2-binary
feedparser
faker
hachoir
pykeyboard
pynewtonmath
aiofiles
ffmpeg-python
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp>=2.2.5
Pillow>=4.2.0
CurrencyConverter
googletrans==4.0.0-rc1
google-trans-new==1.1.9
jikanpy
speedtest-cli
coffeehouse==2.2.4
regex
bleach
wikipedia
markdown2>=2.4.0
telethon
telegraph
heroku3
spamwatch
alphabet_detector
pyrate-limiter
cachetools
ujson
pretty_errors
pygithub
TgCrypto
Pyrogram
PyDictionary
youtube-dl
youtube_search_python
youtube_search
lyrics_extractor
asyncio
cloudscraper
dateparser==1.0.0
pymongo
dnspython
secureme
apscheduler
emoji-country-flag
countryinfo
html2text
bs4
redis
pySmartDL
bing_image_downloader
search_engine_parser
pytz
jikanpy
lyricsgenius
tswift
python-arq
envparse
#profanity (Credits to Julia)
better_profanity
textblob 
nudepy
nekos.py
motor
gpytranslate
fontTools
opencv-python-headless
odmantic
tzlocal
pendulum
